Hide tooltips on form scroll
Because of a longstanding dojo issue we need to make sure to
manually close all visible dijit Tooltip instances.
Thanks to the fact that Tooltip dijit widget operates as a
singleton we can add aspects around open and close to track
active Tooltip instances and hide them on form.scroll event.

Closes #32
Fixes: CMS-15062