Track smart publish command
Track command click and command execute.
isDefaultSelected means all the available content is checked.
noItemSelected means none of the content is checked.
In the case that no sub-content is available to publish, both property
should be true.

Change the return of promise, so that it rejects when cancel is clicked.

customEvents
| where name == "publish"
| sort by timestamp desc
| project CommandType = customDimensions["command-type"],
    ContentType = customDimensions["content-type"],
    customDimensions["smart-publish.is-default-selected"],
    customDimensions["smart-publish.no-item-selected"]
| where CommandType =="smart-command-click" or CommandType == "smart"

CMS-15252