Inline publish promise should return value
InlinePublish command handles `otherwise` promise callback,
and because of that the _ChangeStatusCommand patch will always enter
`then` callback. The return value from InlinePublish `always` callback
is propagated to the promise callbacks chain defined in the patch.
It means that patched command will get `true` if the operation
was successfull and `false` if it failed.

Task: CMS-15295