Set content property to new added ContentArea item
When an item is added to ContentArea, set content property to all the
corresponding items.
Previously we add aspect to dndTarget. It works only when dragging items
to the ContentArea dropzone. Change aspect to addChild, so that it also
works when dragging to the list area.
However there is a protential perf issue: when initializing ContentArea,
request for "getLatestVersions" is sent individually for each item.

Fixes: CMS-15386