Calculate isDirty flag using deep compare
Because of the fact that some dijit controls like
ToggleButton triggers form.change event on load
we have to make sure that we only change the isDirty
flag when the form.value was really changed.
That's why we have to save the initial form value
and compare it with the changed value.