Make sure 'Publish' button is shown/enabled when needed
The publish button should be hidden when the current user
has no publishing access rights.

If he has correct rights then the button should be visible
but disabled.
It should be enabled once the form gets dirty or if there
already was a draft available to be published.

We need to manually trigger the _onModelChange in order
to fetch contentdata from store and calculate if the
InlinePublish command should be available or not.

Fixes: CMS-15715