Fix DDS store
We can index by ParentContentLink so the lookup is fast.
Then we just find the needed block in memory from a small
list of blocks from specific page version.

We need to have Identity prop in DynamicData object so
that we can call `Save` and update existing value
instead of creating duplicates.