Make sure to save/publish nested blocks
We need to skip the check for asset folder when checking
nested content items because an inner block will never
have its own `For this block` but would always reuse the
`parent` folder from either a shared block or a page.
Additionally the infinite loop prevention with `contentReferences`
need to be run later, first we should order to scan the dependencies
and then add the current contentLink so that the same content item
is not processed twice.