Allow the user to configure Labs according to needs
User is now able to control when Quick Edit command
is available. It is possible to turn it on Shared
content if needed (false by default) if for some
reason the customer has many shared blocks and is
not willing to convert them via the admin conversion
tool (introduced in v0.11.0).
It is also possible to additionally filter the list
of types when `Quick Edit` is available via
`IgnoreQuickEditOnBlockTypes` where user can specify
Type[] or `IgnoreQuickEditOnBlockTypeIdentifiers`
with the type identifiers string format.

Example usecase is our `Forms` package.
Form blocks do not work properly in Quick Edit
and should automatically prefer the regular Edit
workflow with `Quick Edit` being disabled at all times.

However, when used as local blocks we should still
maintain its lifecycle in connection to their
direct parent.

Fixes: CMS-21104