Make sure that editing local content notifies the owner
When editing a local content item outside the `Quick Edit`
dialog we need to make sure that its `owner` which is
the first non-local ancestor is notified about the change.
For example if local block is changed we need to find out
its page / shared block parent context and make sure to
trigger the saving event (either create a new draft or
just force to the existing draft).
This effectively causes that parent to scan its direct
dependencies and update the versions in its `Saved`
event (as before).

CMS-21104